requirements.txt
Django==4.2
djangorestframework==3.14.0
Markdown==3.4.3
django-filter==23.1
drf-dynamic-fields==0.4.0
djoser==2.2.0
drf-yasg==1.21.5
